event_driven_poc/microservices/email_validator/requirements.txt
kafka-python==2.0.2
pandas==1.5.3
numpy==1.24.3
minio==7.1.17